requirements.txt
# Core dependencies
fastapi==0.128.8
uvicorn[standard]==0.39.0
pydantic==2.12.5
python-dotenv

# Logging
loguru

# HTTP client
httpx

# Database
motor

# AI/ML
google-genai
